Refactor struct names
Now the VmState struct refers to the World and the VmState before got renamed for Execution.
