+++ src/lib.rs
@@ -7,15 +7,15 @@ mod opcode;
 pub mod output;
 mod precompiles;
 mod ptr_operator;
-pub mod state;
+pub mod execution;
 pub mod store;
 pub mod tracers;
 pub mod utils;
 pub mod value;
 pub mod vm;
 pub use opcode::Opcode;
-pub use state::VMState;
+pub use execution::Execution;
 pub use vm::EraVM;
 mod rollbacks;
-pub mod world;
+pub mod state;
 use zkevm_opcode_defs::Opcode as Variant;
