+++ src/op_handlers/context.rs
@@ -5,7 +5,7 @@ use zkevm_opcode_defs::VmMetaParameters;
 use crate::address_operands::{address_operands_read, address_operands_store};
 use crate::eravm_error::{EraVmError, HeapError};
 use crate::value::TaggedValue;
-use crate::{opcode::Opcode, state::VMState};
+use crate::{opcode::Opcode, execution::Execution};
 
 // consider moving this function to a utils crate
 // taken from matter-labs zk evm implementation
@@ -16,23 +16,23 @@ fn address_to_u256(address: &Address) -> U256 {
     U256::from_big_endian(&buffer)
 }
 
-pub fn this(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn this(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let res =
         TaggedValue::new_raw_integer(address_to_u256(&vm.current_context()?.contract_address));
     address_operands_store(vm, opcode, res)
 }
 
-pub fn caller(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn caller(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let res = TaggedValue::new_raw_integer(address_to_u256(&vm.current_context()?.caller));
     address_operands_store(vm, opcode, res)
 }
 
-pub fn code_address(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn code_address(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let res = TaggedValue::new_raw_integer(address_to_u256(&vm.current_context()?.code_address));
     address_operands_store(vm, opcode, res)
 }
 
-pub fn meta(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn meta(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let res = TaggedValue::new_raw_integer(
         (VmMetaParameters {
             heap_size: vm
@@ -55,28 +55,28 @@ pub fn meta(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
     address_operands_store(vm, opcode, res)
 }
 
-pub fn ergs_left(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn ergs_left(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let res = TaggedValue::new_raw_integer(U256::from(vm.current_frame()?.gas_left.0));
     address_operands_store(vm, opcode, res)
 }
 
-pub fn sp(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn sp(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let sp = vm.current_frame()?.sp;
     address_operands_store(vm, opcode, TaggedValue::new_raw_integer(U256::from(sp)))
 }
 
-pub fn get_context_u128(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn get_context_u128(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let res = TaggedValue::new_raw_integer(U256::from(vm.current_context()?.context_u128));
     address_operands_store(vm, opcode, res)
 }
 
-pub fn set_context_u128(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn set_context_u128(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (src0, _) = address_operands_read(vm, opcode)?;
     vm.register_context_u128 = src0.value.as_u128();
     Ok(())
 }
 
-pub fn increment_tx_number(vm: &mut VMState, _opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn increment_tx_number(vm: &mut Execution, _opcode: &Opcode) -> Result<(), EraVmError> {
     vm.tx_number += 1;
     Ok(())
 }
