+++ src/op_handlers/event.rs
@@ -3,12 +3,12 @@ use zkevm_opcode_defs::ADDRESS_EVENT_WRITER;
 
 use crate::{
     eravm_error::EraVmError,
-    state::VMState,
-    world::{Event, World},
+    execution::Execution,
+    state::{Event, VMState},
     Opcode,
 };
 
-pub fn event(vm: &mut VMState, opcode: &Opcode, world: &mut World) -> Result<(), EraVmError> {
+pub fn event(vm: &mut Execution, opcode: &Opcode, state: &mut VMState) -> Result<(), EraVmError> {
     if vm.current_context()?.contract_address == H160::from_low_u64_be(ADDRESS_EVENT_WRITER as u64)
     {
         let key = vm.get_register(opcode.src0_index).value;
@@ -21,7 +21,7 @@ pub fn event(vm: &mut VMState, opcode: &Opcode, world: &mut World) -> Result<(),
             tx_number: vm.tx_number as u16,
         };
 
-        world.record_event(event);
+        state.record_event(event);
     }
     Ok(())
 }
