+++ src/op_handlers/fat_pointer_read.rs
@@ -3,9 +3,9 @@ use u256::U256;
 use crate::address_operands::address_operands_read;
 use crate::eravm_error::{EraVmError, HeapError, OperandError};
 use crate::value::{FatPointer, TaggedValue};
-use crate::{opcode::Opcode, state::VMState};
+use crate::{opcode::Opcode, execution::Execution};
 
-pub fn fat_pointer_read(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn fat_pointer_read(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (src0, _) = address_operands_read(vm, opcode)?;
     if !src0.is_pointer {
         return Err(OperandError::InvalidSrcNotPointer(opcode.variant).into());
