+++ src/op_handlers/or.rs
@@ -1,9 +1,9 @@
 use crate::address_operands::{address_operands_read, address_operands_store};
 use crate::eravm_error::EraVmError;
 use crate::value::TaggedValue;
-use crate::{opcode::Opcode, state::VMState};
+use crate::{execution::Execution, opcode::Opcode};
 
-pub fn or(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn or(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (src0, src1) = address_operands_read(vm, opcode)?;
 
     let res = src0.value | src1.value;
