+++ src/op_handlers/precompile_call.rs
@@ -13,12 +13,12 @@ use crate::{
         ecrecover::ecrecover_function, keccak256::keccak256_rounds_function,
         secp256r1_verify::secp256r1_verify_function, sha256::sha256_rounds_function,
     },
-    state::VMState,
+    execution::Execution,
     value::TaggedValue,
     Opcode,
 };
 
-pub fn precompile_call(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn precompile_call(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (src0, src1) = address_operands_read(vm, opcode)?;
     let aux_data = PrecompileAuxData::from_u256(src1.value);
