+++ src/op_handlers/ptr_add.rs
@@ -1,12 +1,12 @@
 use crate::{
     eravm_error::{EraVmError, OperandError},
     ptr_operator::{ptr_operands_read, ptr_operands_store},
-    state::VMState,
+    execution::Execution,
     value::FatPointer,
     Opcode,
 };
 
-pub fn ptr_add(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn ptr_add(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (pointer, diff, src0) = ptr_operands_read(vm, opcode)?;
 
     let (new_offset, overflow) = pointer.offset.overflowing_add(diff);
