+++ src/op_handlers/ptr_shrink.rs
@@ -1,12 +1,12 @@
 use crate::{
     eravm_error::{EraVmError, OperandError},
+    execution::Execution,
     ptr_operator::{ptr_operands_read, ptr_operands_store},
-    state::VMState,
     value::FatPointer,
     Opcode,
 };
 
-pub fn ptr_shrink(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn ptr_shrink(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (pointer, diff, src0) = ptr_operands_read(vm, opcode)?;
 
     let (new_len, overflow) = pointer.len.overflowing_sub(diff);
