+++ src/op_handlers/sub.rs
@@ -1,9 +1,9 @@
 use crate::address_operands::{address_operands_read, address_operands_store};
 use crate::eravm_error::EraVmError;
 use crate::value::TaggedValue;
-use crate::{opcode::Opcode, state::VMState};
+use crate::{opcode::Opcode, execution::Execution};
 
-pub fn sub(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn sub(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (src0_t, src1_t) = address_operands_read(vm, opcode)?;
 
     let (src0, src1) = (src0_t.value, src1_t.value);
