+++ src/op_handlers/xor.rs
@@ -1,9 +1,9 @@
 use crate::address_operands::{address_operands_read, address_operands_store};
 use crate::eravm_error::EraVmError;
 use crate::value::TaggedValue;
-use crate::{opcode::Opcode, state::VMState};
+use crate::{opcode::Opcode, execution::Execution};
 
-pub fn xor(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+pub fn xor(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let (src0, src1) = address_operands_read(vm, opcode)?;
 
     let res = src0.value ^ src1.value;
