+++ src/tracers/tracer.rs
@@ -1,7 +1,11 @@
-use crate::{eravm_error::EraVmError, state::VMState, Opcode};
+use crate::{eravm_error::EraVmError, execution::Execution, Opcode};
 
 pub trait Tracer {
-    fn before_execution(&mut self, _opcode: &Opcode, _vm: &mut VMState) -> Result<(), EraVmError> {
+    fn before_execution(
+        &mut self,
+        _opcode: &Opcode,
+        _vm: &mut Execution,
+    ) -> Result<(), EraVmError> {
         Ok(())
     }
 }
