+++ src/vm.rs
@@ -41,11 +41,11 @@ use crate::op_handlers::shift::{rol, ror, shl, shr};
 use crate::op_handlers::sub::sub;
 use crate::op_handlers::unimplemented::unimplemented;
 use crate::op_handlers::xor::xor;
+use crate::state::VMState;
 use crate::store::{ContractStorage, InitialStorage};
 use crate::tracers::blob_saver_tracer::BlobSaverTracer;
 use crate::value::{FatPointer, TaggedValue};
-use crate::world::World;
-use crate::{eravm_error::EraVmError, tracers::tracer::Tracer, VMState};
+use crate::{eravm_error::EraVmError, tracers::tracer::Tracer, Execution};
 use crate::{Opcode, Variant};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -59,7 +59,7 @@ pub enum ExecutionOutput {
 #[derive(Debug)]
 pub struct EraVM {
     pub state: VMState,
-    pub world: World,
+    pub execution: Execution,
 }
 
 pub enum EncodingMode {
@@ -69,13 +69,13 @@ pub enum EncodingMode {
 
 impl EraVM {
     pub fn new(
-        state: VMState,
+        execution: Execution,
         initial_storage: Rc<RefCell<dyn InitialStorage>>,
         contract_storage: Rc<RefCell<dyn ContractStorage>>,
     ) -> Self {
         Self {
-            state,
-            world: World::new(initial_storage, contract_storage),
+            state: VMState::new(initial_storage, contract_storage),
+            execution,
         }
     }
 
@@ -132,17 +132,17 @@ impl EraVM {
     ) -> Result<ExecutionOutput, EraVmError> {
         loop {
             let opcode = match enc_mode {
-                EncodingMode::Testing => self.state.get_opcode_with_test_encode()?,
-                EncodingMode::Production => self.state.get_opcode()?,
+                EncodingMode::Testing => self.execution.get_opcode_with_test_encode()?,
+                EncodingMode::Production => self.execution.get_opcode()?,
             };
             for tracer in tracers.iter_mut() {
-                tracer.before_execution(&opcode, &mut self.state)?;
+                tracer.before_execution(&opcode, &mut self.execution)?;
             }
 
-            let can_execute = self.state.can_execute(&opcode);
+            let can_execute = self.execution.can_execute(&opcode);
 
-            if self.state.decrease_gas(opcode.gas_cost).is_err() || can_execute.is_err() {
-                match inexplicit_panic(&mut self.state, &mut self.world) {
+            if self.execution.decrease_gas(opcode.gas_cost).is_err() || can_execute.is_err() {
+                match inexplicit_panic(&mut self.execution, &mut self.state) {
                     Ok(false) => continue,
                     _ => return Ok(ExecutionOutput::Panic),
                 }
@@ -152,87 +152,95 @@ impl EraVM {
                 let result = match opcode.variant {
                     Variant::Invalid(_) => Err(OpcodeError::InvalidOpCode.into()),
                     Variant::Nop(_) => {
-                        address_operands_read(&mut self.state, &opcode)?;
+                        address_operands_read(&mut self.execution, &opcode)?;
                         address_operands_store(
-                            &mut self.state,
+                            &mut self.execution,
                             &opcode,
                             TaggedValue::new_raw_integer(0.into()),
                         )
                     }
-                    Variant::Add(_) => add(&mut self.state, &opcode),
-                    Variant::Sub(_) => sub(&mut self.state, &opcode),
-                    Variant::Jump(_) => jump(&mut self.state, &opcode),
-                    Variant::Mul(_) => mul(&mut self.state, &opcode),
-                    Variant::Div(_) => div(&mut self.state, &opcode),
+                    Variant::Add(_) => add(&mut self.execution, &opcode),
+                    Variant::Sub(_) => sub(&mut self.execution, &opcode),
+                    Variant::Jump(_) => jump(&mut self.execution, &opcode),
+                    Variant::Mul(_) => mul(&mut self.execution, &opcode),
+                    Variant::Div(_) => div(&mut self.execution, &opcode),
                     Variant::Context(context_variant) => match context_variant {
-                        ContextOpcode::AuxMutating0 => unimplemented(&mut self.state, &opcode),
-                        ContextOpcode::Caller => caller(&mut self.state, &opcode),
-                        ContextOpcode::CodeAddress => code_address(&mut self.state, &opcode),
-                        ContextOpcode::ErgsLeft => ergs_left(&mut self.state, &opcode),
-                        ContextOpcode::GetContextU128 => get_context_u128(&mut self.state, &opcode),
-                        ContextOpcode::IncrementTxNumber => {
-                            increment_tx_number(&mut self.state, &opcode)
+                        ContextOpcode::AuxMutating0 => unimplemented(&mut self.execution, &opcode),
+                        ContextOpcode::Caller => caller(&mut self.execution, &opcode),
+                        ContextOpcode::CodeAddress => code_address(&mut self.execution, &opcode),
+                        ContextOpcode::ErgsLeft => ergs_left(&mut self.execution, &opcode),
+                        ContextOpcode::GetContextU128 => {
+                            get_context_u128(&mut self.execution, &opcode)
                         }
-                        ContextOpcode::Meta => meta(&mut self.state, &opcode),
-                        ContextOpcode::SetContextU128 => set_context_u128(&mut self.state, &opcode),
-                        ContextOpcode::Sp => sp(&mut self.state, &opcode),
-                        ContextOpcode::This => this(&mut self.state, &opcode),
+                        ContextOpcode::IncrementTxNumber => {
+                            increment_tx_number(&mut self.execution, &opcode)
+                        }
+                        ContextOpcode::Meta => meta(&mut self.execution, &opcode),
+                        ContextOpcode::SetContextU128 => {
+                            set_context_u128(&mut self.execution, &opcode)
+                        }
+                        ContextOpcode::Sp => sp(&mut self.execution, &opcode),
+                        ContextOpcode::This => this(&mut self.execution, &opcode),
                     },
                     Variant::Shift(shift_variant) => match shift_variant {
-                        ShiftOpcode::Shl => shl(&mut self.state, &opcode),
-                        ShiftOpcode::Shr => shr(&mut self.state, &opcode),
-                        ShiftOpcode::Rol => rol(&mut self.state, &opcode),
-                        ShiftOpcode::Ror => ror(&mut self.state, &opcode),
+                        ShiftOpcode::Shl => shl(&mut self.execution, &opcode),
+                        ShiftOpcode::Shr => shr(&mut self.execution, &opcode),
+                        ShiftOpcode::Rol => rol(&mut self.execution, &opcode),
+                        ShiftOpcode::Ror => ror(&mut self.execution, &opcode),
                     },
                     Variant::Binop(binop) => match binop {
-                        BinopOpcode::Xor => xor(&mut self.state, &opcode),
-                        BinopOpcode::And => and(&mut self.state, &opcode),
-                        BinopOpcode::Or => or(&mut self.state, &opcode),
+                        BinopOpcode::Xor => xor(&mut self.execution, &opcode),
+                        BinopOpcode::And => and(&mut self.execution, &opcode),
+                        BinopOpcode::Or => or(&mut self.execution, &opcode),
                     },
                     Variant::Ptr(ptr_variant) => match ptr_variant {
-                        PtrOpcode::Add => ptr_add(&mut self.state, &opcode),
-                        PtrOpcode::Sub => ptr_sub(&mut self.state, &opcode),
-                        PtrOpcode::Pack => ptr_pack(&mut self.state, &opcode),
-                        PtrOpcode::Shrink => ptr_shrink(&mut self.state, &opcode),
+                        PtrOpcode::Add => ptr_add(&mut self.execution, &opcode),
+                        PtrOpcode::Sub => ptr_sub(&mut self.execution, &opcode),
+                        PtrOpcode::Pack => ptr_pack(&mut self.execution, &opcode),
+                        PtrOpcode::Shrink => ptr_shrink(&mut self.execution, &opcode),
                     },
-                    Variant::NearCall(_) => near_call(&mut self.state, &opcode, &self.world),
+                    Variant::NearCall(_) => near_call(&mut self.execution, &opcode, &self.state),
                     Variant::Log(log_variant) => match log_variant {
                         LogOpcode::StorageRead => {
-                            storage_read(&mut self.state, &opcode, &self.world)
+                            storage_read(&mut self.execution, &opcode, &self.state)
                         }
                         LogOpcode::StorageWrite => {
-                            storage_write(&mut self.state, &opcode, &mut self.world)
+                            storage_write(&mut self.execution, &opcode, &mut self.state)
                         }
                         LogOpcode::ToL1Message => {
-                            add_l2_to_l1_message(&mut self.state, &opcode, &mut self.world)
+                            add_l2_to_l1_message(&mut self.execution, &opcode, &mut self.state)
                         }
-                        LogOpcode::PrecompileCall => precompile_call(&mut self.state, &opcode),
-                        LogOpcode::Event => event(&mut self.state, &opcode, &mut self.world),
+                        LogOpcode::PrecompileCall => precompile_call(&mut self.execution, &opcode),
+                        LogOpcode::Event => event(&mut self.execution, &opcode, &mut self.state),
                         LogOpcode::Decommit => {
-                            opcode_decommit(&mut self.state, &opcode, &mut self.world)
+                            opcode_decommit(&mut self.execution, &opcode, &mut self.state)
                         }
                         LogOpcode::TransientStorageRead => {
-                            transient_storage_read(&mut self.state, &opcode, &self.world)
+                            transient_storage_read(&mut self.execution, &opcode, &self.state)
                         }
                         LogOpcode::TransientStorageWrite => {
-                            transient_storage_write(&mut self.state, &opcode, &mut self.world)
+                            transient_storage_write(&mut self.execution, &opcode, &mut self.state)
                         }
                     },
                     Variant::FarCall(far_call_variant) => {
-                        let res =
-                            far_call(&mut self.state, &opcode, &far_call_variant, &mut self.world);
+                        let res = far_call(
+                            &mut self.execution,
+                            &opcode,
+                            &far_call_variant,
+                            &mut self.state,
+                        );
                         if res.is_err() {
-                            panic_from_far_call(&mut self.state, &opcode)?;
+                            panic_from_far_call(&mut self.execution, &opcode)?;
                             continue;
                         }
                         Ok(())
                     }
                     Variant::Ret(ret_variant) => match ret_variant {
                         RetOpcode::Ok => {
-                            match ret(&mut self.state, &opcode, &mut self.world, ret_variant) {
+                            match ret(&mut self.execution, &opcode, &mut self.state, ret_variant) {
                                 Ok(should_break) => {
                                     if should_break {
-                                        let result = retrieve_result(&mut self.state)?;
+                                        let result = retrieve_result(&mut self.execution)?;
                                         return Ok(ExecutionOutput::Ok(result));
                                     }
                                     Ok(())
@@ -241,10 +249,10 @@ impl EraVM {
                             }
                         }
                         RetOpcode::Revert => {
-                            match ret(&mut self.state, &opcode, &mut self.world, ret_variant) {
+                            match ret(&mut self.execution, &opcode, &mut self.state, ret_variant) {
                                 Ok(should_break) => {
                                     if should_break {
-                                        let result = retrieve_result(&mut self.state)?;
+                                        let result = retrieve_result(&mut self.execution)?;
                                         return Ok(ExecutionOutput::Revert(result));
                                     }
                                     Ok(())
@@ -253,7 +261,7 @@ impl EraVM {
                             }
                         }
                         RetOpcode::Panic => {
-                            match ret(&mut self.state, &opcode, &mut self.world, ret_variant) {
+                            match ret(&mut self.execution, &opcode, &mut self.state, ret_variant) {
                                 Ok(should_break) => {
                                     if should_break {
                                         return Ok(ExecutionOutput::Panic);
@@ -265,9 +273,9 @@ impl EraVM {
                         }
                     },
                     Variant::UMA(uma_variant) => match uma_variant {
-                        UMAOpcode::HeapRead => heap_read(&mut self.state, &opcode),
+                        UMAOpcode::HeapRead => heap_read(&mut self.execution, &opcode),
                         UMAOpcode::HeapWrite => {
-                            let result = heap_write(&mut self.state, &opcode);
+                            let result = heap_write(&mut self.execution, &opcode);
                             match result {
                                 exec_hook @ Ok(ExecutionOutput::SuspendedOnHook { .. }) => {
                                     return exec_hook
@@ -277,11 +285,11 @@ impl EraVM {
                             }
                         }
 
-                        UMAOpcode::AuxHeapRead => aux_heap_read(&mut self.state, &opcode),
-                        UMAOpcode::AuxHeapWrite => aux_heap_write(&mut self.state, &opcode),
-                        UMAOpcode::FatPointerRead => fat_pointer_read(&mut self.state, &opcode),
-                        UMAOpcode::StaticMemoryRead => unimplemented(&mut self.state, &opcode),
-                        UMAOpcode::StaticMemoryWrite => unimplemented(&mut self.state, &opcode),
+                        UMAOpcode::AuxHeapRead => aux_heap_read(&mut self.execution, &opcode),
+                        UMAOpcode::AuxHeapWrite => aux_heap_write(&mut self.execution, &opcode),
+                        UMAOpcode::FatPointerRead => fat_pointer_read(&mut self.execution, &opcode),
+                        UMAOpcode::StaticMemoryRead => unimplemented(&mut self.execution, &opcode),
+                        UMAOpcode::StaticMemoryWrite => unimplemented(&mut self.execution, &opcode),
                     },
                 };
                 if let Err(err) = result {
@@ -289,21 +297,21 @@ impl EraVM {
                         return Ok(ExecutionOutput::Panic);
                     }
 
-                    match inexplicit_panic(&mut self.state, &mut self.world) {
+                    match inexplicit_panic(&mut self.execution, &mut self.state) {
                         Ok(false) => continue,
                         _ => return Ok(ExecutionOutput::Panic),
                     }
                 }
-                set_pc(&mut self.state, &opcode)?;
+                set_pc(&mut self.execution, &opcode)?;
             } else {
-                self.state.current_frame_mut()?.pc += 1;
+                self.execution.current_frame_mut()?.pc += 1;
             }
         }
     }
 }
 
 // Sets the next PC according to the next opcode
-fn set_pc(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
+fn set_pc(vm: &mut Execution, opcode: &Opcode) -> Result<(), EraVmError> {
     let current_pc = vm.current_frame()?.pc;
 
     vm.current_frame_mut()?.pc = match opcode.variant {
@@ -317,7 +325,7 @@ fn set_pc(vm: &mut VMState, opcode: &Opcode) -> Result<(), EraVmError> {
     Ok(())
 }
 
-fn retrieve_result(vm: &mut VMState) -> Result<Vec<u8>, EraVmError> {
+fn retrieve_result(vm: &mut Execution) -> Result<Vec<u8>, EraVmError> {
     let fat_pointer_src0 = FatPointer::decode(vm.get_register(1).value);
     let range = fat_pointer_src0.start..fat_pointer_src0.start + fat_pointer_src0.len;
     let mut result: Vec<u8> = vec![0; range.len()];
